video: tegra: host: Make client_managed field wider
Allow client managed sync points to occupy range 0-63 instead of the
old 0-31.

Reviewed-on: http://git-master/r/205930

+++ drivers/video/tegra/host/host1x/host1x.h
@@ -37,7 +37,7 @@ struct host1x_device_info {
 	int		nb_channels;	/* host1x: num channels supported */
 	int		nb_pts; 	/* host1x: num syncpoints supported */
 	int		nb_bases;	/* host1x: num syncpoints supported */
-	u32		client_managed; /* host1x: client managed syncpts */
+	u64		client_managed; /* host1x: client managed syncpts */
 	int		nb_mlocks;	/* host1x: number of mlocks */
 	const char	**syncpt_names;	/* names of sync points */
 };
